Fix plugin crashing on uninstallation
