+++ includes/wp-admin.php
@@ -119,17 +119,3 @@ function admin_bar_ticket_count( \WP_Admin_Bar $admin_bar ) {
 }
 
 add_action( 'admin_bar_menu', 'ucare\admin_bar_ticket_count', 80 );
-
-
-function add_plugin_action_links( $links ) {
-
-    if( get_option( Options::DEV_MODE ) !== 'on' ) {
-        $links['deactivate'] = '<span id="feedback-prompt">' . $links['deactivate'] . '</span>';
-    }
-
-    $menu_page = menu_page_url( 'uc-settings', false );
-
-    return array_merge( array( 'settings' => '<a href="' . $menu_page . '">' . __( 'Settings', 'ucare' ) . '</a>' ), $links );
-}
-
-add_action( 'plugin_action_links_' . basename(), 'ucare\add_plugin_action_links' );
